Only calculate subset of Rp and Vp required
Rp was a 3D matrix too large to fit in memory,
but only a subset was used for plotting, so we only calculate that.

Vp was also a 3D matrix too large for memory,
and was an intermediate matrix used to calculate some totals.
We now only calculate a 2D matrix,
and calculate the sum in a loop instead of storing each result
in a separate page.